refactor: move user post related components into one folder
